Introduce query/search feature; add LQL
See DESIGN_NOTES_NAMESPACE_AND_QUERY.md for a full brief on which
features are being currently introduced. This commit introduces
a new query and metadata endpoint (see swagger.html) and the
lockd query language (LQL, exportable package lql). The client
mutations and the rudimentary query scan feature now use LQL for
building the query selector and for constructing/mutating json
state objects. Next up is a Cassandra-like search feature,
secondary index, manifests, segments, compaction, etc with the
goal to have a reasonably performant search of state objects within
a namespace. Attributes was also introduced as metadata in order
to exclude objects (if you want to) from appearing in queries,
which is useful for lockd internals like the diagnostic objects,
but also to keep some objects "hidden" from searches that you
don't want part of the query returns (perhaps workflow objects
and not searchable entities like "orders", etc). The search
feature only applies to state objects and not queues (this
includes stateful dequeues, those states are or will not be
searchable).